Add FFT-based convolution extractor
The file FFT.java contains and implementation of FFT for power-of-two
vectors and, using that, inverse-FFT and generalizations to 2D.
Everything is based on fft1D(), so changing that to use something else
than the naive FFT implementation that is included would a) make things
faster and b) allow non-power-of-2 lengths.

FFT.java contains also convolution, that can be used with pre-calculated
FFTs. Does not strip the results, as it has no information about
padding, so the whole 0-padded circular convolution is returned.

FFTConvolutionExtractor implementes Extractor using the FFT class.
When setFeatures() is called, the FFT's of the features are calculated
and stored. When call() is called, the FFT for that data vector is
calculated, and then it is convolved with each pre-calculated feature
FFT, and the result is stripped to contain only valid cells. The results
are stored in a single vector, with each feature being row-major and
featues are one after another, in the same order as they were given to
setFeatures(). Preprocessing, if PreProcessZCA is used, is taken care
of.

FFTConvolutionTest.java test the FFT operations and the
FFTConvolutionExtractor with both vector and matrix data/features.
Preprocessing is not tested yet.